Refs #24682: Storyblok - move loading placeholder into the `BaseImage` component
