+++ components/storyblok/Image.vue
@@ -10,22 +10,15 @@
       @close="lightboxIndexValue = null"
     />
 
-    <div
-      class="_image-wrapper"
-    >
-      <div class="_placeholder" />
-
-      <SfImage
-        class="_image"
-        :src="itemData.image.filename"
-        :mobile-src="itemData.mobile_image.filename"
-        :alt="itemData.alt_tag"
-        :title="itemData.title_tag"
-        @load.capture="onLoad"
-        @click="launchLightbox"
-        :width="itemData.width"
-      />
-    </div>
+    <SfImage
+      class="_image"
+      :src="itemData.image.filename"
+      :mobile-src="itemData.mobile_image.filename"
+      :alt="itemData.alt_tag"
+      :title="itemData.title_tag"
+      @click="launchLightbox"
+      :width="itemData.width"
+    />
   </div>
 </template>
 
@@ -39,7 +32,6 @@ import { Blok } from 'src/modules/vsf-storyblok-module/components';
 import SfImage from './BaseImage.vue';
 import ImageData from './interfaces/image-data.interface';
 import LightboxItemValue from './interfaces/lightbox-item-value.interface';
-import generatePlaceholderStyles from './generate-placeholder-styles';
 
 const SCREEN_WIDTH_BREAKPOINT = 768;
 
@@ -51,28 +43,15 @@ export default Blok.extend({
   },
   data () {
     return {
-      lightboxIndexValue: null as number | null,
-      isLoaded: false
+      lightboxIndexValue: null as number | null
     }
   },
   computed: {
     itemData (): ImageData {
       return this.item as ImageData;
     },
-    extraCssClasses (): string[] {
-      return !this.isLoaded ? ['-loading'] : [];
-    },
     extraStyles (): Record<string, string> {
-      const styles = generatePlaceholderStyles(
-        this.itemData.image.filename,
-        this.itemData.mobile_image.filename,
-        'image-block-height'
-      );
-
-      styles['--image-width'] = '100%';
-      if (this.itemData.width) {
-        styles['--image-width'] = this.itemData.width;
-      }
+      const styles: Record<string, string> = {};
 
       if (this.itemData.has_lightbox) {
         styles['cursor'] = 'pointer';
@@ -109,52 +88,14 @@ export default Blok.extend({
       }
 
       this.lightboxIndexValue = 0;
-    },
-    onLoad (): void{
-      this.isLoaded = true;
     }
   }
 })
 </script>
 
 <style lang="scss">
-@import "~@storefront-ui/shared/styles/helpers/breakpoints";
-
 .storyblok-image {
   text-align: center;
   font-size: 0;
-
-  ._image-wrapper {
-    display: inline-block;
-    position: relative;
-    width: var(--image-width, 100%);
-
-    ._image {
-      position: absolute;
-      left: 0;
-      top: 0;
-      width: 100%;
-      height: 100%;
-    }
-
-    ._placeholder {
-      background-color: none;
-      padding-top: var(--image-block-height-mobile, var(--image-block-height, 0));
-    }
-  }
-
-  &.-loading {
-    ._placeholder {
-      background-color: #fafafa;
-    }
-  }
-
-  @media (min-width: $tablet-min) {
-    ._image-wrapper {
-      ._placeholder {
-        padding-top: var(--image-block-height, 0);
-      }
-    }
-  }
 }
 </style>
